gemini20-all/requirements.txt
google-genai
numpy
SpeechRecognition
websockets
markdown
pygments
chromadb
pypdf
python-dotenv